gpt_mastermind.py
```python
import argparse
from gpt_integration import GPTIntegration
from code_analysis import CodeAnalysis
from suggestions import Suggestions
from error_detection import ErrorDetection
from readability_enhancements import ReadabilityEnhancements
from efficiency_optimization import EfficiencyOptimization
from documentation_generation import DocumentationGeneration
from security import Security
from error_handling import ErrorHandling

class GPTMastermind:
    def __init__(self, api_key):
        self.gpt = GPTIntegration(api_key)
        self.code_analysis = CodeAnalysis(api_key)
        self.suggestions = Suggestions(api_key)
        self.error_detection = ErrorDetection(api_key)
        self.readability_enhancements = ReadabilityEnhancements(api_key)
        self.efficiency_optimization = EfficiencyOptimization(api_key)
        self.documentation_generation = DocumentationGeneration(api_key)
        self.security = Security(api_key)
        self.error_handling = ErrorHandling(api_key)

    def analyze_code(self, code):
        analysis_results = self.code_analysis.analyze_code(code)
        return analysis_results

    def suggest_improvements(self, code):
        suggestions = self.suggestions.suggest_improvements(code)
        return suggestions

    def detect_errors(self, code):
        errors = self.error_detection.detect_errors(code)
        return errors

    def enhance_readability(self, code):
        readability = self.readability_enhancements.enhance_readability(code)
        return readability

    def optimize_efficiency(self, code):
        efficiency = self.efficiency_optimization.optimize_efficiency(code)
        return efficiency

    def generate_documentation(self, code):
        documentation = self.documentation_generation.generate_documentation(code)
        return documentation

    def handle_errors(self, code):
        error_handling = self.error_handling.handle_errors(code)
        return error_handling

    def secure_code(self, code):
        secure_code = self.security.secure_code(code)
        return secure_code

def main():
    parser = argparse.ArgumentParser(description='GPT-Mastermind: AI-Generated Python Code Analyzer')
    parser.add_argument('code', type=str, help='Python code to analyze and enhance')
    args = parser.parse_args()

    api_key = 'your_openai_api_key'  # Replace with your OpenAI API key

    gpt_mastermind = GPTMastermind(api_key)

    code = args.code

    analysis_results = gpt_mastermind.analyze_code(code)
    print(f"Code Analysis Results: {analysis_results}")

    suggestions = gpt_mastermind.suggest_improvements(code)
    print(f"Suggestions for Improvements: {suggestions}")

    errors = gpt_mastermind.detect_errors(code)
    print(f"Detected Errors: {errors}")

    readability = gpt_mastermind.enhance_readability(code)
    print(f"Readability Enhancements: {readability}")

    efficiency = gpt_mastermind.optimize_efficiency(code)
    print(f"Efficiency Optimizations: {efficiency}")

    documentation = gpt_mastermind.generate_documentation(code)
    print(f"Generated Documentation: {documentation}")

    error_handling = gpt_mastermind.handle_errors(code)
    print(f"Error Handling: {error_handling}")

    secure_code = gpt_mastermind.secure_code(code)
    print(f"Secured Code: {secure_code}")

if __name__ == "__main__":
    main()
```
