test_gpt_mastermind.py
```python
import unittest
from gpt_integration import GPTIntegration
from code_analysis import CodeAnalysis
from suggestions import Suggestions
from error_detection import ErrorDetection
from readability_enhancements import ReadabilityEnhancements
from efficiency_optimization import EfficiencyOptimization

class TestGPTMastermind(unittest.TestCase):
    def setUp(self):
        self.api_key = "your_openai_api_key"
        self.gpt = GPTIntegration(self.api_key)
        self.code_analysis = CodeAnalysis(self.api_key)
        self.suggestions = Suggestions(self.api_key)
        self.error_detection = ErrorDetection(self.api_key)
        self.readability_enhancements = ReadabilityEnhancements(self.api_key)
        self.efficiency_optimization = EfficiencyOptimization(self.api_key)
        self.test_code = "def add(a, b):\n    return a + b"

    def test_gpt_integration(self):
        response = self.gpt.generate_code_suggestions(self.test_code)
        self.assertIsNotNone(response)

    def test_code_analysis(self):
        analysis = self.code_analysis.analyze_code(self.test_code)
        self.assertIsNotNone(analysis)
        self.assertIn("structure_analysis", analysis)
        self.assertIn("readability_analysis", analysis)
        self.assertIn("efficiency_analysis", analysis)

    def test_suggestions(self):
        improvements = self.suggestions.suggest_improvements(self.test_code)
        self.assertIsNotNone(improvements)

    def test_error_detection(self):
        errors = self.error_detection.detect_errors(self.test_code)
        self.assertIsNotNone(errors)

    def test_readability_enhancements(self):
        enhancements = self.readability_enhancements.enhance_readability(self.test_code)
        self.assertIsNotNone(enhancements)

    def test_efficiency_optimization(self):
        optimization = self.efficiency_optimization.optimize_efficiency(self.test_code)
        self.assertIsNotNone(optimization)

if __name__ == "__main__":
    unittest.main()
```
